detection-engine/requirements.txt
kafka-python==2.0.2
qdrant-client==1.6.0
numpy==1.25.2
